requirements.txt
Telethon == 1.34.0
python-dotenv == 1.0.1
python-telegram-bot == 20.8
nlp-id == 0.1.15.0
nltk
scikit-learn
